requirements.txt
alembic==1.18.4
apispec==6.9.0
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.1.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
webargs==8.7.1
Werkzeug==3.1.6
